+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
-gem 'rspec', '~> 2.4'
+gem 'rspec', '~> 3.1' 
 require 'rspec'
 require 'deployable/patch/version'
+
 
 include Deployable::Patch