+++ ChangeLog.md
@@ -1,4 +1,9 @@
 ### 0.1.0 / 2014-10-19
 
-* Initial release:
+* 0.2.0
+  Add hash any_keys all_keys + specs
+  Add specs for not_nil 
+  Add ancestor classes and class helper methods
 
+* Initial release:
+ 0.1.0